Add files via upload
I come to you with my newest invention - discoverer of free IP addresses! All you have to do is input a network id in a x.y.z format, input numbers from 1 to 254, how many free addresses you want it to discover and a name of a list file you want it to create, press enter and sit back as it works!

Important notes:
- do NOT input spaces nor extensions in your list name, it adds the extension on its own;
- if you input a same list name twice, it will CONTINUE with the input without erasing previous data, but it will add the date and time of your previous session;
- if you give it a range of 10 addresses, but you tell it to look for 30 free addresses, it will ping only the 10 addresses you told it to;

Please do note that this script is a work-in-progress, its content is subject to change and you are free to use it, shape it and give any feedback/suggestions!